Moves hooks folder from common-component to utils
