+++ react/src/utils/hooks/useInitCsrf.js
